laboratory_work_3/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14
djoser>=2.2
